breast cancer/requirements.txt
flask
pandas
scikit-learn
joblib
pyyaml
numpy
